Move data models to ciscosparkapi.models
Separate the data models from the API modules by moving them into their own modules and package.
