+++ ciscosparkapi/api/roles.py
@@ -1,23 +1,25 @@
 # -*- coding: utf-8 -*-
-"""Cisco Spark Roles API wrapper.
-
-Classes:
-    Role: Models a Spark Role JSON object as a native Python object.
-    RolesAPI: Wraps the Cisco Spark Roles API and exposes the API as native
-        Python methods that return native Python objects.
-
-"""
+"""Cisco Spark Roles API."""
 
 
-# Use future for Python v2 and v3 compatibility
 from __future__ import (
     absolute_import,
     division,
     print_function,
     unicode_literals,
 )
+
 from builtins import *
+
 from past.builtins import basestring
+
+from ..generator_containers import generator_container
+from ..models.role import Role
+from ..restsession import RestSession
+from ..utils import (
+    check_type,
+    dict_from_items_with_values,
+)
 
 
 __author__ = "Chris Lunsford"
@@ -26,43 +28,8 @@ __copyright__ = "Copyright (c) 2016-2018 Cisco and/or its affiliates."
 __license__ = "MIT"
 
 
-from ..generator_containers import generator_container
-from ..restsession import RestSession
-from ..sparkdata import SparkData
-from ..utils import (
-    check_type,
-    dict_from_items_with_values,
-)
-
-
-class Role(SparkData):
-    """Model a Spark Role JSON object as a native Python object."""
-
-    def __init__(self, json):
-        """Initialize a new Role data object from a dictionary or JSON string.
-
-        Args:
-            json(dict, basestring): Input dictionary or JSON string.
-
-        Raises:
-            TypeError: If the input object is not a dictionary or string.
-
-        """
-        super(Role, self).__init__(json)
-
-    @property
-    def id(self):
-        """The unique ID for the Role."""
-        return self._json_data.get('id')
-
-    @property
-    def name(self):
-        """The name of the Role."""
-        return self._json_data.get('name')
-
-
 class RolesAPI(object):
-    """Cisco Spark Roles API wrapper.
+    """Cisco Spark Roles API.
 
     Wraps the Cisco Spark Roles API and exposes the API as native Python
     methods that return native Python objects.
