+++ ciscosparkapi/api/teams.py
@@ -1,23 +1,25 @@
 # -*- coding: utf-8 -*-
-"""Cisco Spark Teams API wrapper.
-
-Classes:
-    Team: Models a Spark 'team' JSON object as a native Python object.
-    TeamsAPI: Wraps the Cisco Spark Teams API and exposes the API as native
-        Python methods that return native Python objects.
-
-"""
+"""Cisco Spark Teams API."""
 
 
-# Use future for Python v2 and v3 compatibility
 from __future__ import (
     absolute_import,
     division,
     print_function,
     unicode_literals,
 )
+
 from builtins import *
+
 from past.builtins import basestring
+
+from ..generator_containers import generator_container
+from ..models.team import Team
+from ..restsession import RestSession
+from ..utils import (
+    check_type,
+    dict_from_items_with_values,
+)
 
 
 __author__ = "Chris Lunsford"
@@ -26,53 +28,8 @@ __copyright__ = "Copyright (c) 2016-2018 Cisco and/or its affiliates."
 __license__ = "MIT"
 
 
-from ..generator_containers import generator_container
-from ..restsession import RestSession
-from ..sparkdata import SparkData
-from ..utils import (
-    check_type,
-    dict_from_items_with_values,
-)
-
-
-class Team(SparkData):
-    """Model a Spark 'team' JSON object as a native Python object."""
-
-    def __init__(self, json):
-        """Initialize a new Team data object from a dictionary or JSON string.
-
-        Args:
-            json(dict, basestring): Input dictionary or JSON object.
-
-        Raises:
-            TypeError: If the input object is not a dictionary or string.
-
-        """
-        super(Team, self).__init__(json)
-
-    @property
-    def id(self):
-        """The team's unique ID."""
-        return self._json_data.get('id')
-
-    @property
-    def name(self):
-        """A user-friendly name for the team."""
-        return self._json_data.get('name')
-
-    @property
-    def created(self):
-        """The date and time the team was created."""
-        return self._json_data.get('created')
-
-    @property
-    def creatorId(self):
-        """The ID of the person who created the team."""
-        return self._json_data.get('creatorId')
-
-
 class TeamsAPI(object):
-    """Cisco Spark Teams API wrapper.
+    """Cisco Spark Teams API.
 
     Wraps the Cisco Spark Teams API and exposes the API as native Python
     methods that return native Python objects.
