+++ examples/people.py
@@ -9,7 +9,6 @@ variable set to run this script.
 """
 
 
-# Use future for Python v2 and v3 compatibility
 from __future__ import (
     absolute_import,
     division,
