+++ tests/api/test_messages.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """pytest Messages functions, fixtures and tests."""
+import ciscosparkapi.models.message
 
 
 __author__ = "Chris Lunsford"
@@ -41,7 +42,8 @@ def list_messages(api, roomId, **query_parameters):
 
 
 def is_valid_message(obj):
-    return isinstance(obj, ciscosparkapi.Message) and obj.id is not None
+    return isinstance(obj,
+                      ciscosparkapi.models.message.Message) and obj.id is not None
 
 
 def are_valid_messages(iterable):
